Continuted Data Cleansing Work
Assessment of missing values and imputations of values
Checked ELEV for logical imputation measures but all records in locations proved problematic. May use similarity based on latitide and longitude but as this only accounts for 3% of cases the decision was made to remove.
Further work in relation to correcting data types and dealing with remainder of NA - This part of the process can be considered complete